+++ CHANGELOG.md
@@ -25,7 +25,8 @@ Features:
 * Write job-executions (ad-hoc and scheduled) to audit-log [#49](https://github.com/O-X-L/ansible-webui/issues/49)
 * Fixed issue with ansible-runner usage [ansible-runner #1398](https://github.com/ansible/ansible-runner/issues/1398) - changed back to official runner
 * Security: Clean temporary job-credentials before starting job
-* API & Frontend to add and manage SSH-hostkeys [#6](https://github.com/O-X-L/ansible-webui/issues/6), [#77](https://github.com/O-X-L/ansible-webui/issues/77)
+* API & Frontend to add and manage SSH-hostkeys [#77](https://github.com/O-X-L/ansible-webui/issues/77), [#6](https://github.com/O-X-L/ansible-webui/issues/6)
+* API & Form-integration for inventory-list (limit-field suggestions) [#78](https://github.com/O-X-L/ansible-webui/issues/78), [#29](https://github.com/O-X-L/ansible-webui/issues/29)
 
 ----